Avances_22-11(21:19)

+++ Laberinto/src/Nodo.java
@@ -1,16 +1,16 @@
 public class Nodo {
 	
 	private long id;
-	private int costo;
+	private float costo;
 	private int[] id_estado;
 	private long id_padre;
 	private int profundidad;
-	private int heuristica;
+	private double heuristica;
 	private float valor;
 	private String accion;
 
 	
-	public Nodo(long id, int costo, int[] id_estado, long id_padre, String accion, int profundidad, int heuristica, float valor) {
+	public Nodo(long id, float costo, int[] id_estado, long id_padre, String accion, int profundidad, double heuristica, float valor) {
 		this.id = id;
 		this.costo = costo;
 		this.id_estado = id_estado;
@@ -29,11 +29,11 @@ public class Nodo {
 		this.id = id;
 	}
 
-	public int getCosto() {
+	public float getCosto() {
 		return costo;
 	}
 
-	public void setCosto(int costo) {
+	public void setCosto(float costo) {
 		this.costo = costo;
 	}
 
@@ -61,11 +61,11 @@ public class Nodo {
 		this.profundidad = profundidad;
 	}
 
-	public int getHeuristica() {
+	public double getHeuristica() {
 		return heuristica;
 	}
 
-	public void setHeuristica(int heuristica) {
+	public void setHeuristica(double heuristica) {
 		this.heuristica = heuristica;
 	}
